Refactor components to replace useConnectWallet with usePrivyAddress hook for improved address management
