+++ components/Header/Header.tsx
@@ -7,6 +7,7 @@ import usePrivyAddress from "@/hooks/usePrivyAddress";
 import Button from "../Button";
 import LoginButton from "../LoginButton";
 import Tooltip from "../ui/Tooltip";
+import usePrivyAddress from "@/hooks/usePrivyAddress";
 
 const Header = () => {
   const { push } = useRouter();
